refactored treedoc files
