+++ treedoc/__main__.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-This module is the entrypoint when a user calls
+This module is the entrypoint when a user calls:
 $ python -m treedoc [object]
 """
 
-from treedoc.main import main
+from treedoc.main import CLI_entrypoint
 
 if __name__ == "__main__":
-    main()
+    CLI_entrypoint()
